server-tray-app/requirements.txt
fastapi
uvicorn
pycaw
comtypes
pyautogui
psutil
pystray
Pillow
winrt-Windows.Devices.Radios
winrt-Windows.Foundation.Collections
